+++ docs/DECISIONS.md
@@ -97,6 +97,7 @@
 - Review diff presentation is built by `WorkspaceReviewSurfaceBuilder`. `WorkspaceSurface.swift` supplies tool cards and thread events, while the builder owns latest successful `host.git.diff` selection, diff parsing, file/line review comment attachment, timestamp ordering, and line-kind filtering.
 - Context pressure presentation is built by `WorkspaceContextBannerBuilder`. `WorkspaceSurface.swift` supplies the selected thread, while the builder owns empty-thread hiding, context estimate calculation, warning thresholds, full-context copy, and New/Fork/Compact command surfaces.
 - Transcript projection is built by `WorkspaceTranscriptSurfaceBuilder`. `WorkspaceModel` supplies selected threads and enriches project execution context after projection, while the builder owns visible message filtering, feedback reduction, tool-card construction, safety-review cards, artifact projection, and message/tool timeline interleaving.
+- Transcript surface contracts live in `QuillCodeTranscriptSurface.swift`. `TranscriptSurface`, timeline rows, `ContextBannerSurface`, `MessageSurface`, and `ComposerSurface` own empty-state copy, timeline IDs, compatibility decoding, message accessibility labels, sendability, and slash suggestions; `WorkspaceSurface.swift` should only carry the aggregate transcript/context/composer payloads and assemble them through the focused builders.
 - Runtime issue diagnostics are generated from current runtime state, not hardcoded per error. Settings shows the API base URL, authentication mode, key state, model, agent status, rate-limit metadata when present, and a redacted last-error snippet when present. Secret-like `sk-...` and bearer tokens are redacted before they reach the surface.
 - Context pressure recovery is a thread action, not only a warning. The first compaction path creates a new `Compact:` thread that preserves project/mode/model/instructions/memories, adds a bounded assistant summary of older turns, and keeps the latest actionable turn. This keeps behavior deterministic while leaving room for a TrustedRouter-backed compactor later.
 - Thread archive is reversible and distinct from delete. Archived chats move to an Archived sidebar section, remain searchable, and can be unarchived back into the active thread list. Delete removes the thread from memory and the JSON thread store. Rename uses an explicit title field in the native UI, while duplicate creates an unpinned active `Copy:` thread that preserves the transcript and event audit.
